add get_invoices_summary schema json
