+++ dynatademand/schemas/request/query/get_invoices_summary.json
@@ -1,1 +1,17 @@
-{}
+{
+  "type": "object",
+  "properties": {
+    "startDate": {
+      "type": "string",
+      "description": "The format is YY-MM-DD"
+    },
+    "endDate": {
+      "type": "string",
+      "description": "The format is YY-MM-DD"
+    },
+    "extProjectId": {
+      "type": "string"
+    }
+  },
+  "required": []
+}
