+++ tests/test_line_items.py
@@ -1,5 +1,5 @@
-from __future__ import unicode_literals, print_function
 # encoding: utf-8
+from __future__ import unicode_literals, print_function
 
 import json
 import unittest
